refactor(client): move game components into the game screen folder
#52
