--- packages/client/src/components/GameContainer/index.ts
@@ -1,3 +1,0 @@
-import QuizConainer from './QuizContainer';
-
-export { QuizConainer };
