--- packages/client/src/components/GameMainScreen/index.ts
@@ -1,3 +1,0 @@
-import QuizMainScreen from './QuizMainScreen';
-
-export { QuizMainScreen };
