+++ packages/client/src/components/ScoreBoard/Scoreboard/ScoreboardStyle.tsx
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    position: 'absolute',
+    left: 0,
+    top: 0,
+    right: 0,
+    bottom: 0,
+    overflow: 'hidden',
+    paddingRight: 20,
+  },
+});
+
+export default styles;
