+++ packages/client/src/components/ScoreBoard/Scoreboard/index.ts
@@ -1,0 +1,3 @@
+import Scoreboard from './Scoreboard';
+
+export default Scoreboard;
