+++ packages/client/src/screens/Room/GameEnd/ScoreBoard/FullScoreboard/GameEndFullScoreboard.tsx
@@ -1,5 +1,5 @@
 import Card from '@Src/components/Card';
-import { ScoreBoardContent } from '@Src/components/ScoreBoard';
+import Scoreboard from '@Src/components/ScoreBoard';
 import { DisplayPlayer } from '@Src/global/playerInfoState';
 import React from 'react';
 import { View } from 'react-native';
@@ -15,7 +15,7 @@ export default function GameEndFullScoreBoard({ players }: GameEndFullScoreBoard
     <View style={styles.container}>
       <Card style={styles.card}>
         <View style={styles.grow}>
-          <ScoreBoardContent players={players} />
+          <Scoreboard players={players} />
         </View>
       </Card>
     </View>
