+++ packages/client/src/screens/Room/GameInProgress/Answer/GameInProgressAnswerStyle.tsx
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  answers: {
+    textAlign: 'center',
+    fontWeight: 'bold',
+  },
+  container: {
+    paddingTop: 20,
+    shadowOffset: undefined,
+    paddingVertical: 20,
+    borderRadius: 10,
+  },
+});
+
+export default styles;
