+++ packages/client/src/screens/Room/GameInProgress/Answer/index.ts
@@ -1,0 +1,3 @@
+import GameInProgressAnswer from './GameInProgressAnswer';
+
+export default GameInProgressAnswer;
