+++ packages/client/src/screens/Room/GameInProgress/GameInPrgress/index.ts
@@ -1,0 +1,3 @@
+import GameInProgess from './GameInProgress';
+
+export default GameInProgess;
