+++ packages/client/src/screens/Room/GameInProgress/Question/index.ts
@@ -1,0 +1,3 @@
+import GameInProgressQuestion from './GameInProgressQuestion';
+
+export default GameInProgressQuestion;
