+++ packages/client/src/screens/Room/GameInProgress/RoundCounter/index.ts
@@ -1,0 +1,3 @@
+import RoundCounter from './RoundCounter';
+
+export default RoundCounter;
