+++ packages/client/src/screens/Room/GameInProgress/Scoreboard/index.ts
@@ -1,0 +1,3 @@
+import GameInProgressScoreboard from './GameInProgressScoreBoard';
+
+export default GameInProgressScoreboard;
