+++ packages/client/src/screens/Room/GameInProgress/index.ts
@@ -1,0 +1,3 @@
+import GameInProgress from './GameInPrgress';
+
+export default GameInProgress;
