+++ packages/client/src/screens/Room/Room/Room.tsx
@@ -1,32 +1,22 @@
-/* import { QuizConainer } from '@Src/components/GameContainer';
-import getEnv from '@Src/constant/index';
-import socketState from '@Src/global/socket';
-import userState from '@Src/global/userState'; */
+import userState from '@Src/global/userState';
 import { HomeNavigatorProps } from '@Src/typings/navigation';
-/* import { useFocusEffect } from '@react-navigation/native'; */
+import useSocketConnect from '@Src/utils/hooks/socketConnect';
+import { useSocketListener } from '@Src/utils/hooks/socketListener';
 import React from 'react';
-/* import { useRecoilValue, useSetRecoilState } from 'recoil';
-import io from 'socket.io-client'; */
+import { useRecoilValue } from 'recoil';
+import { RoomStatus } from 'shared/src/enums/Room';
 
 import GameEnd from '../GameEnd';
+import GameInProgess from '../GameInProgress/GameInPrgress';
 
 export default function Room({ route }: HomeNavigatorProps<'Room'>) {
-  /* const setSocket = useSetRecoilState(socketState);
   const { username } = useRecoilValue(userState);
+  useSocketConnect(route.params.id, { pseudo: username });
+  const status: { status: RoomStatus } = useSocketListener('status', RoomStatus.Waiting);
 
-  useFocusEffect(
-    React.useCallback(() => {
-      if (!username) return;
-      const socket = io(getEnv().backendUrl + route.params.id, {
-        query: {
-          pseudo: username,
-        },
-        reconnectionAttempts: 3,
-      });
-      setSocket(socket);
-      return () => socket.close();
-    }, [username])
-  ); */
-
-  return <GameEnd />;
+  if (status.status === RoomStatus.Ended) {
+    return <GameEnd />;
+  } else {
+    return <GameInProgess status={status.status} />;
+  }
 }
